requirements.txt
Django==3.1
django-allauth==0.42.0
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
humanize==2.6.0
Pillow==7.2.0
six==1.15.0
Jinja2==2.11.1
django-quill-editor
python-decouple
gunicorn
whitenoise
psycopg2
boto3
django-storages